+++ hlib/python/hlib/frontend/__init__.py
@@ -1,3 +1,4 @@
 from . import relay_parser
 from . import test
 from . import utils
+from . import relay_attr
