+++ hlib/python/hlib/frontend/test/test_keras.py
@@ -25,7 +25,6 @@ def verify_keras_frontend(keras_model, need_trans_before=True,need_trans_after=T
     def get_hcl_output(xs, dtype='float32'):
         shape_dict = {name: x.shape for (name, x) in zip(keras_model.input_names, xs)}
         #return relay_front.from_keras(keras_model, shape_dict)
-        print(shape_dict)
         return get_relay_model(keras_model, shape_dict, 'keras')
 
     def to_channels_first(arr):
@@ -42,7 +41,6 @@ def verify_keras_frontend(keras_model, need_trans_before=True,need_trans_after=T
 
     xs = [np.random.randint(size=shape, low=1, high=10).astype(dtype) for shape in in_shapes]
     keras_out = get_keras_output(xs,dtype)
-    print(len(keras_out))
     inputs = [to_channels_first(x) for x in xs] if need_trans_before else xs
     f,params = get_hcl_output(inputs,dtype)
     out = []
@@ -61,8 +59,6 @@ def verify_keras_frontend(keras_model, need_trans_before=True,need_trans_after=T
         print(_out.shape)
     f(*inputs,*params,*out)
     if(isinstance(keras_out,(tuple,list))):
-        #print(out)
-        #print(keras_out)
         for i in range(len(keras_out)):
             if(need_trans_after):
                 h_out = out[i].asnumpy()
@@ -141,7 +137,6 @@ def merge_conv_test():
             out = merge_func([x, y])
         else:
             out = merge_func([x, y, z])
-    #out = keras.layers.Add()([x,y])
     keras_model = keras.models.Model(data, out)
     verify_keras_frontend(keras_model,True,True)
 
@@ -295,14 +290,12 @@ def conv_code_test():
         strides=strides,dilation=dilation)
         return hlib.op.nn.bias_add(i_0,bias,axis=axis)
     s = hcl.create_schedule([input_1,param_1,param_2],func)
-    print(hcl.lower(s))
     f = hcl.build(s)
     _in = hcl.asarray(np.random.randint(10,size=(1,3,3,3)))
     filt = hcl.asarray(np.random.randint(10,size=(3,3,3,3)))
     bias = hcl.asarray(np.random.randint(10,size=(3,)))
     out = hcl.asarray(np.zeros((1,3,3,3)))
     f(_in,filt,bias,out)
-    print(out.asnumpy())
 
 def test_forward_multi_inputs():
     data1 = keras.layers.Input(shape=(32, 32, 3))
@@ -378,7 +371,6 @@ def test_forward_conv():
                   #keras.layers.Conv2DTranspose(filters=10, kernel_size=(3, 3), padding='valid'),
                   keras.layers.SeparableConv2D(filters=10, kernel_size=(3, 3), padding='same')]
     for conv_func in conv_funcs:
-        print(conv_func)
         x = conv_func(data)
         keras_model = keras.models.Model(data, x)
         verify_keras_frontend(keras_model,True,True)
@@ -425,7 +417,6 @@ def test_forward_activations():
     for act_func in act_funcs:
         x = act_func(data)
         keras_model = keras.models.Model(data, x)
-        print(keras_model.summary())
         verify_keras_frontend(keras_model,False,False)
 
 
@@ -469,55 +460,52 @@ def test_forward_vgg16():
 def test_forward_xception():
     keras_model = keras.applications.Xception(include_top=True, weights='imagenet',
         input_shape=(299, 299, 3), classes=1000)
-    print(keras_model.summary())
     verify_keras_frontend(keras_model,True,False)
 
 
 def test_forward_resnet50():
     keras_model = keras.applications.ResNet50(include_top=True, weights='imagenet',
         input_shape=(224, 224, 3), classes=1000)
-    print(keras_model.summary())
     verify_keras_frontend(keras_model,True,False)
 
 
 def test_forward_mobilenet():
     keras_model = keras.applications.MobileNet(include_top=True, weights='imagenet',
         input_shape=(224, 224, 3), classes=1000)
-    print(keras_model.summary())
     verify_keras_frontend(keras_model,True,False,'float64')
 
 if __name__ == "__main__":
-    #test_for_paper()
-    #merge_test((2,2))
-    #merge_test((10,7,4))
-    #merge_2_test((3,3))
-    #pooling_test((32,32,16))
-    #pooling_test((32,16,32))
-    #pooling_test((16,32,32))
+    test_for_paper()
+    merge_test((2,2))
+    merge_test((10,7,4))
+    merge_2_test((3,3))
+    pooling_test((32,32,16))
+    pooling_test((32,16,32))
+    pooling_test((16,32,32))
     #dot_test()
-    #sequential_test()
-    #rnn_test()
-    #reshape_test()
-    #simple_pool_test()
-    #merge_and_pool_test((16,8,4))
-    #merge_and_pool_test((8,8,8))
-    #merge_out_tup_test((4,4,4))
-    #merge_just_conv_test()
-    #test_forward_conv()
-    #test_depthwise_conv()
-    #test_separable_conv()
-    #test_forward_multi_inputs()
-    #test_forward_multi_outputs()
-    #test_reuse_layers()
-    #conv_code_test()
-    #merge_conv_test()
-    #dense_test()
-    #test_forward_activations()
-    #cifar10_test()
+    sequential_test()
+    rnn_test()
+    reshape_test()
+    simple_pool_test()
+    merge_and_pool_test((16,8,4))
+    merge_and_pool_test((8,8,8))
+    merge_out_tup_test((4,4,4))
+    merge_just_conv_test()
+    test_forward_conv()
+    test_depthwise_conv()
+    test_separable_conv()
+    test_forward_multi_inputs()
+    test_forward_multi_outputs()
+    test_reuse_layers()
+    conv_code_test()
+    merge_conv_test()
+    dense_test()
+    test_forward_activations()
+    cifar10_test()
     test_forward_vgg16()
     test_forward_xception()
     test_forward_resnet50()
-    #batch_norm_test((4,4),1)
-    #test_forward_mobilenet()
-    #test_multiple_reuse()
+    batch_norm_test((4,4),1)
+    test_forward_mobilenet()
+    test_multiple_reuse()
     print("All Passed!")
